==> nsiegel/sprint_1/02-scrape_reviews.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Load-Libraries-and-Data" data-toc-modified-id="Load-Libraries-and-Data-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Load Libraries and Data</a></span></li><li><span><a href="#Extract-Review-Text" data-toc-modified-id="Extract-Review-Text-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Extract Review Text</a></span></li><li><span><a href="#Save-to-File" data-toc-modified-id="Save-to-File-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Save to File</a></span></li></ul></div>

# Notebook Purpose
This notebook uses the reddit API to scrape review text and append them to the reddit review file.

## Load Libraries and Data

import praw
import pandas as pd
import re

import requests
import time
import sys
import pdb

rdb = pd.read_parquet('data/redditwhiskeydatabase.parquet')

Client secrets for reddit are loaded here. You'll need to provide your own secrets.py which provides the following fields. You'll need to sign up for a reddit api to get this.:

- clientid
- clientsecret
- clientusername
- clientpassword

import secrets

## Extract Review Text

def extract_reviews(dataframe,re_client_id,re_client_secret,\
                    re_username,re_pw):
        """Takes Reddit review submission ids, request \
        submissions from PRAW, search for the right body text,\
        and saves it to the dataframe""" 
        #timer/counter
        start_time = int(time.time())
        count = 0
        total = dataframe.shape[0]
        
        #get the Review URL addresses as a series
        rev_addr = dataframe['sub_id']

        #empty list of reviews
        reviews = []
        
        #initiate PRAW
        reddit = praw.Reddit(client_id= re_client_id,\
                     client_secret = re_client_secret,\
                     username = re_username, \
                             password = re_pw,user_agent= 'v1' )
        
        
        
        #iterate through the submission ids to extract review texts
        for s_id in rev_addr:
            review = ''
            try:
                sub = reddit.submission(id = s_id)
                s_author = sub.author.name
            except:
                reviews.append('error')
                continue
            #sometimes the reviews are in the 
            #submission text instead of comments
            #add any text in this field if not empty
            sub_text = sub.selftext
            if len(sub_text) > 0:
                review = review + sub_text + ' \n'

            # parse through the submission's comments 
            # and extract all of authors' comments
            for comment in sub.comments:
                try:
                    if comment.author.name == s_author:
                        review = review + comment.body + '\n'
                except:
                    continue
            
            reviews.append(review)
            
            #counter/timer
            end_time = int(time.time())
            min_spent = round(((end_time-start_time)/60),2)
            remaining_time = round((min_spent/max(count,1))*max((total-count),1),2)
            
            
            count_str = 'extracted ' + str(count) + '/' + str(total) + \
                    ' reviews in total ' + str(min_spent) + ' minutes' + \
                    ' estimated time remaining is ' + str(remaining_time) + ' minutes'
            sys.stdout.write("\r{}".format(count_str))
            count += 1
            
        dataframe['review'] = reviews
        #save to csv, also return the pandas dataframe
        dataframe.to_csv("db_reviews.csv")
        return dataframe
        

reviews = extract_reviews(rdb,clientid,clientsecret,\
                    clientusername,clientpassword)

reviews.shape

## Save to File

reviews.to_parquet('db_reviews.parquet')

reviews.tail()
