==> nsiegel/sprint_1/03-split_review_text.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Load-Libraries-and-Data" data-toc-modified-id="Load-Libraries-and-Data-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Load Libraries and Data</a></span></li><li><span><a href="#Extract-Review-Text" data-toc-modified-id="Extract-Review-Text-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Extract Review Text</a></span></li><li><span><a href="#Save-to-File" data-toc-modified-id="Save-to-File-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Save to File</a></span></li></ul></div>

# Notebook Purpose
This notebook splits up the review text into the categories we need: Nose, Taste, Finish.
This works by splitting the text into lines and using some regular expressions.

## Load Libraries and Data

import pandas as pd
import re

import multiprocessing as mp

rdb = pd.read_parquet('data/db_reviews.parquet')

Take a look at a review:


rdb.iloc[0]['review']

Can look through this by splitting on new lines:

def extract_categories(text,name):
    nose_re = '.*nose.{,4}[:-]|\*nose\*?|[\*\s]n[:-]'
    taste_re = '.*taste.{,4}[:-]|\*taste\*?|\*palate\*|.*palate.{,6}[:-]|[\*\s]t[:-]|[\*\s]p[:-]'
    finish_re = '.*finish.{,4}[:-]|\*finish\*?|[\*\s]f[:-]'

    # Initialize Collections
    review_categories = ['','','']
    review_count = [0,0,0]
    
    review = re.split("\n+", text.lower())
    
    for i, line in enumerate(review):
        # for each line split
        nose = re.findall(nose_re, line)
        taste = re.findall(taste_re, line)
        finish = re.findall(finish_re, line)
              
        searchresults = [nose, taste, finish]
        for idx, result in enumerate(searchresults):
            if result:
                review_count[idx] += 1
                # if the category title is on a line before the review:
                if len(line) < 15 and i<len(review)-1:
                    review_categories[idx] += review[i+1]
                else:
                # otherwise it's on the same line
                    review_categories[idx] += line

    return {'nose': re.sub("(nose)|\*|:|>",'', review_categories[0]),
            'taste': re.sub("(taste)|(palate)|\*|:|>",'', review_categories[1]),
            'finish':re.sub("(finish)|\*|:|>",'', review_categories[2])}
                

# Function to multiprocess an entire dataframe
def extract_categories_dataframe(df, columnname):
    # create dataframe to hold results
    global results
    results = pd.DataFrame(columns=['reviewID','results'])
    
    # select only the column we want and make unique to save some time
    reviewlist = df[['reviewID','whisky', columnname]].drop_duplicates()
    pool = mp.Pool(mp.cpu_count())
    
    # call function for each name
    for review in reviewlist.itertuples():
        pool.apply_async(extract_categories_row, args=(review.reviewID, review.review, review.whisky), callback=collect_result)
    pool.close()
    pool.join()
    
    # break out dictionary from results
    results = pd.concat([results.drop(['results'], axis=1), results['results'].apply(pd.Series)], axis=1)
    
    # join back on original dataframe
    return (df.set_index('reviewID')
              .join(results.set_index('reviewID'))
              .reset_index()
              .rename({'index':'reviewID'}, axis='columns')
           )
    
# Function to be ran in multiprocess on each item
def extract_categories_row(reviewID, text, name):
    newitem = {}
    newitem['reviewID'] = reviewID
    newitem['results'] = extract_categories(text, name)
    return newitem
    
# Function to collect results from multiprocess
def collect_result(result):
    global results
    results = results.append(result,ignore_index = True)

## Extract Review Text

Filter out reviews without any text

results = None
rdb = pd.read_parquet('data/db_reviews.parquet').reset_index().rename({'index':'reviewID'}, axis='columns')
rdb = rdb[rdb['review'] != '']
rdb = extract_categories_dataframe(rdb, 'review')

## Save to File

rdb.to_parquet('data/db_reviews_split.parquet')
